services/telegram_bot/requirements.txt
python-telegram-bot
google-genai
python-decouple
